pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cocomind"
version = "0.1.0"
description = "AI-assisted CRPF Tender Evaluation and Eligibility Analysis Platform"
readme = "CLAUDE.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "CocoMind Team"}]

dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.12",

    # Jobs
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",

    # Document processing
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "Pillow>=10.4.0",
    "pdf2image>=1.17.0",
    "python-magic-bin>=0.4.14",

    # OCR
    "azure-ai-formrecognizer>=3.3.0",

    # LLM
    "anthropic>=0.39.0",

    # Embeddings & retrieval
    "sentence-transformers>=3.3.0",
    "chromadb>=0.5.0",
    "whoosh>=2.7.4",

    # PII redaction
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",

    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",

    # Reporting
    "reportlab>=4.2.0",

    # Crypto & signing
    "cryptography>=43.0.0",

    # Observability
    "structlog>=24.4.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "prometheus-fastapi-instrumentator>=7.0.0",

    # Fuzzy matching
    "jellyfish>=1.1.0",

    # Jinja2 for prompts
    "Jinja2>=3.1.0",

    # UI
    "streamlit>=1.39.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "hypothesis>=6.112.0",
    "mypy>=1.11.0",
    "ruff>=0.7.0",
    "import-linter>=2.1",
    "httpx>=0.27.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "RET", "SIM"]
ignore = ["S101", "E501", "B008", "S105", "S112"]  # allow assert in tests, ignore line limit, FastAPI File(), false-positive passwords, try-except-pass

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "slow: marks tests as slow",
    "e2e: end-to-end tests",
    "red_team: prompt injection tests",
]

[tool.importlinter]
root_packages = ["src"]
include_external_packages = true

[[tool.importlinter.contracts]]
name = "Rule engine must not import LLM SDKs"
type = "forbidden"
source_modules = ["src.engine"]
forbidden_modules = ["anthropic", "openai", "langchain", "langchain_core", "langchain_anthropic"]
